Rename Appointment to AppointmentsDayView
Given reacts nested usage, when combining smaller related
components in a single file, name the file after the root
component
