+++ appointments/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { AppointmentsDayView } from './Appointment';
+import { AppointmentsDayView } from './AppointmentsDayView';
 import { sampleAppointments } from './sampleData';
 
 ReactDOM.render(
